requirements.txt
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
xgboost>=1.5.0
plotly>=5.18.0
selenium>=4.1.0
beautifulsoup4>=4.9.0
dash>=2.14.0
scipy>=1.7.0
statsmodels>=0.13.0
